Uploaded example code using joystick API
js_linux.py is a python script that uses the open-source joystick API for linux to talk to the USB Lynxmotion controller. It does not contain any motor code! Originally this script just checked for events logged in the joystick event queue, and printed them, it has now been updated to operate as a pseudo-continuous loop. 
Note that the underlying code is still event-based (the structure storing the joystick state is only updated on events), but we can eg  poll this structure regularly to make continuous steering decisions.